Added __name__ to unary_f in nary_to_unary to give clearer error messages in check_grads. Also factored out common logic for adding name/doc metadata
